openstack: add CI job for provider network
This patch will add a new CI job that will run on demand to test the
deployment of OpenShift on OpenStack using a Provider Network, which
means a network that already exists and is managed by the OpenStack
operators.
It's an extension to what we test with BYON (Bring Your Own Network) but
in this scenario we don't create the network for each CI job, we use one
that already exists and managed by the OpenStack operator.

Changes (by chronology):

* provision-machinesubnet: Pre-create the VIPs for API and Ingress by
  creating Neutron ports and later they'll be removed. This is useful
  so we can "reserve" the IPs, in case another job is also using the
  provider network (remember, this is a shared network). It also writes
  the necessary files in SHARED_DIR.

* provision-bastionproxy: when using a provider network, the bastion
  will be plugged on both the bastion subnet (created by CI tools) and
  the provider network (managed by OpenStack admins and exists
  already). Also add a default for BASTION_USER and improve logging by
  printing the server console when the bastion fails to start.

* openstack-conf-generateconfig: in this step we cleanup the ports that
  we created earlier to reserve them. Now we don't need it anymore
  because the installer will re-create them during the install process.

* deprovision-machinesubnet: Only remove the router if it exists, and
  remove leftover ports for API & Ingress, if they were reserved but
  something failed before the openshift-install execution.

Note: this work will also be useful to test the deployment of a cluster
spread across multiple availability zones, that will be connected
through a provider network.

+++ ci-operator/step-registry/openstack/deprovision/machinesubnet/openstack-deprovision-machinesubnet-commands.sh
@@ -14,12 +14,26 @@ fi
 export OS_CLIENT_CONFIG_FILE=${SHARED_DIR}/clouds.yaml
 
 CLUSTER_NAME=$(<"${SHARED_DIR}"/CLUSTER_NAME)
-NET_ID=$(<"${SHARED_DIR}"/MACHINESSUBNET_NET_ID)
-SUBNET_ID=$(<"${SHARED_DIR}"/MACHINESSUBNET_SUBNET_ID)
-ROUTER_ID=$(<"${SHARED_DIR}"/MACHINESSUBNET_ROUTER_ID)
 
 >&2 echo "Starting the network cleanup for cluster name $CLUSTER_NAME"
-openstack router remove subnet ${ROUTER_ID} ${SUBNET_ID} || >&2 echo "Failed to delete subnet ${SUBNET_ID} from router ${ROUTER_ID}"
-openstack router delete ${ROUTER_ID} || >&2 echo "Failed to delete router ${ROUTER_ID}"
-openstack network delete ${NET_ID} || >&2 echo "Failed to delete network ${NET_ID}"
+if [[ -f ${SHARED_DIR}"/BASTIONSUBNET_ROUTER_ID" ]]; then
+    ROUTER_ID=$(<"${SHARED_DIR}"/BASTIONSUBNET_ROUTER_ID)
+    if [[ -f ${SHARED_DIR}"/BASTIONSUBNET_SUBNET_ID" ]]; then
+        SUBNET_ID=$(<"${SHARED_DIR}"/BASTIONSUBNET_SUBNET_ID)
+        openstack router remove subnet ${ROUTER_ID} ${SUBNET_ID} || >&2 echo "Failed to delete subnet ${SUBNET_ID} from router ${ROUTER_ID}"
+    fi
+    openstack router delete ${ROUTER_ID} || >&2 echo "Failed to delete router ${ROUTER_ID}"
+fi
+if [[ -f ${SHARED_DIR}"/BASTIONSUBNET_NET_ID" ]]; then
+    NET_ID=$(<"${SHARED_DIR}"/BASTIONSUBNET_NET_ID)
+    openstack network delete ${NET_ID} || >&2 echo "Failed to delete network ${NET_ID}"
+fi
+
+for p in api ingress; do
+  if openstack port show ${CLUSTER_NAME}-${p} >/dev/null; then
+    echo "Leftover port exists: ${CLUSTER_NAME}-${p}, probably due to a failure before install, removing it"
+    openstack port delete ${CLUSTER_NAME}-${p} || >&2 echo "Failed to delete port ${CLUSTER_NAME}-${p}"
+  fi
+done
+
 >&2 echo 'Cleanup done.'
